% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{AsymptoticConfidenceInterval}
\alias{AsymptoticConfidenceInterval}
\title{No export.}
\usage{
AsymptoticConfidenceInterval(
  thetaEst,
  n_sample,
  Cov,
  qLaw = qnorm,
  level = 0.95,
  type,
  ...
)
}
\description{
No export.
}
