% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charCGMY}
\alias{charCGMY}
\title{Function title}
\usage{
charCGMY(t, C, G, M, Y)
}
\arguments{
\item{t}{A gap holder.}

\item{C}{A real number > 0.}

\item{G, M}{A  real number > 0.}

\item{Y}{A real number between 0 and 2.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
charCGMY(1,1,1,1,0.5)
charCGMY(10,1,1,1,0.5)

}
