% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charCTS}
\alias{charCTS}
\title{Characteristic function of the classical tempered stable (CTS) distribution}
\usage{
charCTS(
  t,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A positive integer.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}
}
\value{
The result of spectral positive tempered stable process
}
\description{
Let  \code{X ~ CTS(alpha,deltap, deltam, lambdap, lambdam, mu)}.
The characteristic function is given by (Kuechler & Tappe 2013).
}
\details{
\deqn{\varphi_{CTS}(t;\theta):=
\mathbb{E}_{\theta}\left[
\mathrm{e}^{\mathrm{i}tX}\right]=
\exp\left(\mathrm{i}t\mu+\delta_+\Gamma(-\alpha)
\left((\lambda_+-\mathrm{i}t)^{\alpha}-\lambda_+^{\alpha}+
\mathrm{i}t\alpha\lambda_+^{\alpha-1}\right)+\delta_-\Gamma(-\alpha)
\left((\lambda_-+\mathrm{i}t)^{\alpha}-\lambda_-^{\alpha}-\mathrm{i}t\alpha
\lambda_-^{\alpha-1}\right)
\right)}
}
\examples{
charCTS(2,0.5,1,1.5,2,2.5,3)
charCTS(2,1.5,200,150,300,2500,1,NULL)

}
\references{
Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws';

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distributions and
processes' \url{https://doi.org/10.1016/j.spa.2013.06.012}
}
