% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charNTS}
\alias{charNTS}
\title{Characteristic function of the normal tempered stable (NTS) distribution}
\usage{
charNTS(
  t,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A numeric vector of quantile.}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{Any real number.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}
}
\value{
The result of spectral positive tempered stable process
}
\description{
We can obtain the NTS distribution by tempering a stable distribution.
The corresponding tempering function can be found in Rachev et al. (2011).
}
\details{
\deqn{\varphi_{NTS}(t;\theta)=E\left[\mathrm{e}^{\mathrm{i}tZ}\right]= \exp
\left(\mathrm{i}t\mu+\delta\Gamma(-\alpha)\left((\lambda-\mathrm{i}t
\beta+t^2/2)^{\alpha}-\lambda^{\alpha}\right)\right)
}
}
\examples{
x <- seq(-10,10,0.25)
y <- charNTS(x,0.5,1,1,1,0)
plot(x,y)

}
\references{
Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws'

Rachev, S., Kim, Y., Bianchi, M. & Fabozzi, F. (2011), 'Financial Models with
Levy Processes and Volatility Clustering'
\url{https://books.google.de/books?id=XKvUUrcS_twC}
}
