% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{chartocdf}
\alias{chartocdf}
\title{Cumulative distribution function from characteristic function}
\usage{
chartocdf(q, N, m, s, char, ...)
}
\arguments{
\item{q}{A numeric vector of positive quantile.}

\item{N}{A gap holder.}

\item{m}{A gap holder.}

\item{s}{A gap holder.}

\item{char}{Choose a characteristic function from package. Either [charSTS],
[charCTS], or [charNTS].}

\item{...}{Use this for parameters of the relevant characteristic function.}
}
\value{
As \code{q} is a numeric vector, the return value is also a numeric
vector.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
chartocdf(seq(0,15,0.25),61,1,1,charNTS, alpha=0.5, beta = 1, delta = 1, lambda = 1,
 mu = 1)

}
