% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charTSS}
\alias{charTSS}
\title{Characteristic function of the tempered stable subordinator}
\usage{
charTSS(t, alpha = NULL, delta = NULL, lambda = NULL, theta = NULL)
}
\arguments{
\item{t}{A numeric vector of quantile.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}
}
\value{
The result of spectral positive tempered stable process.
}
\description{
When \code{alpha} is a real number between 0 and 1, the characterisic
function of the tempered stable distribution (\code{alpha}, \code{delta},
\code{lambda}) can be simulated exactly through acceptance-rejection
sampling.
For whole derivation and meaning of single letters see also Kawai et. Masuda
(2011).
Basically, the derivation starts from a one-dimensional stable distribution
which is later added with a centered and totally positively skewed tempered
stable (Levy) process.
}
\details{
\deqn{ \varphi_{TSS}(t;\theta):=\mathbb{E}_{\theta}\left[
\mathrm{e}^{\mathrm{i}tY}\right]= \exp\left(\delta\Gamma(-\alpha)
\left((\lambda-\mathrm{i}t)^{\alpha}-\lambda^{\alpha}\right)\right)}
}
\examples{
x <- seq(-10,10,0.25)
y <- charTSS(x,0.5,1,1)
plot(x,y)

}
\references{
Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws'

Reiichiro K. & Hiroki M. (2011), 'On simulation of tempered stable random
variates' \url{https://doi.org/10.1016/j.cam.2010.12.014}
}
