% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{CheckParametersRange_CTS}
\alias{CheckParametersRange_CTS}
\title{Function title}
\usage{
CheckParametersRange_CTS(theta)
}
\arguments{
\item{theta}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
Not sure, if export is really necessary.
}
\details{
Gap holder for details.
}
