% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{ComputeMCSimForTempered_parallel}
\alias{ComputeMCSimForTempered_parallel}
\title{Function title}
\usage{
ComputeMCSimForTempered_parallel(
  MCparam,
  thetaT,
  size,
  TemperedType = c("Classic", "Subordinator", "Normal", "CGMY"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  eps,
  ...
)
}
\arguments{
\item{MCparam}{A gap holder.}

\item{thetaT}{A gap holder.}

\item{size}{A gap holder.}

\item{TemperedType}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{Estimfct}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{HandleError}{A Boolean. \code{TRUE} by default.}

\item{eps}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
