% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{cospi2}
\alias{cospi2}
\title{tan(pi*x), exact for integer x}
\usage{
cospi2(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of values tan(pi*x)

numeric vector of values cos(pi/2 *x)
}
\description{
tan(pi*x), exact for integer x

cos(pi/2 * x), exact for integer x
}
\author{
Martin Maechler
tanpi <- function(x) tan(pi * (x \%\% 1))

Martin Maechler
}
