% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dCGMY}
\alias{dCGMY}
\title{Title}
\usage{
dCGMY(x, C, G, M, Y, dens_method = "FFT", a = -20, b = 20, nf = 2048, ...)
}
\arguments{
\item{x}{A gap holder.}

\item{C}{A real number > 0.}

\item{G, M}{A  real number > 0.}

\item{Y}{A real number between 0 and 2.}

\item{dens_method}{A gap holder. "FFT" by default.}

\item{a}{A gap holder, if \code{dens_method == "FFT"}. -20
by default.}

\item{b}{A gap holder, if \code{dens_method == "FFT"}. 20
by default.}

\item{nf}{A gap holder.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
dCGMY(1,1,1,1,0.5)
dCGMY(1,1,1,1,0.5,"", -2, 2, 2^4)

}
