% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dCTS}
\alias{dCTS}
\title{Title}
\usage{
dCTS(
  x,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  dens_method = "FFT",
  a = -20,
  b = 20,
  nf = 2048,
  ...
)
}
\arguments{
\item{x}{A gap holder.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{dens_method}{A gap holder.}

\item{a}{A gap holder, if \code{dens_method == "FFT"}. -20
by default.}

\item{b}{A gap holder, if \code{dens_method == "FFT"}. 20
by default.}

\item{nf}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
dCTS(1,0.6,1,1,1,1,1,NULL,"FFT",-20,20,2048)
dCTS(1,0.6,1,1,1,1,1,NULL,"Conv") <-- Warnings will show up.

}
