% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{deleteCheckPoint}
\alias{deleteCheckPoint}
\title{No export.}
\usage{
deleteCheckPoint(
  ParameterMatrix,
  TemperedType,
  Estimfct,
  nab,
  npar,
  nSS,
  MCparam,
  ...
)
}
\description{
No export.
}
