% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dNTS}
\alias{dNTS}
\title{Title}
\usage{
dNTS(
  x,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  dens_method = "FFT",
  a = -20,
  b = 20,
  nf = 2048,
  ...
)
}
\arguments{
\item{x}{A gap holder.}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{A gap holder}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{dens_method}{Currently useless param, as it does nothing. "FFT" by
default.}

\item{a}{A gap holder. -20 by default.}

\item{b}{A gap holder. 20 by default.}

\item{nf}{A gap holder. 2048 by default.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
dNTS(1,0.8,1,1,1,1)
dNTS(0.5,0.5,20,20,20,20, a = -2000, b = 2000, nf = 8192)

}
