% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dSTS}
\alias{dSTS}
\title{PDF TS subordinator by Stable subordinator}
\usage{
dSTS(x, alpha = NULL, delta = NULL, lambda = NULL, theta = NULL)
}
\arguments{
\item{x}{A positive integer.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}
}
\value{
As \code{x} is a numeric vector, the return value is also a numeric
vector
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
dSTS(1000,0.5,1,0.3)
dSTS(1,0.5,1,0.3)

}
