% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dTSS}
\alias{dTSS}
\title{Density function of the tempered stable subordinator (TSS) distribution}
\usage{
dTSS(x, alpha = NULL, delta = NULL, lambda = NULL, theta = NULL)
}
\arguments{
\item{x}{A numeric vector of quantile.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}
}
\value{
As \code{x} is a numeric vector, the return value is also a numeric
vector.
}
\description{
The probability density function of tempered stable distributions is
generally not available in closed form. However, many software packages (like
the \code{stabledist} package) have fast computation routines based on series
or integral representations. Combining such series representations with the
density function of the TSS distribution, we obtain a series representation
for the TSS distribution.
}
\details{
\deqn{f_{TSS}(y;\theta)=\mathrm{e}^{-\lambda y-\lambda^{\alpha}\delta
\Gamma(-\alpha)}\frac{-1}{\pi}\sum_{k=1}^{\infty}\frac{(-1)^k}{k!}
\Gamma(1+\alpha k)\Gamma(1-\alpha)^k\left(\frac{\delta}{\alpha}\right)^ky^
{-(1+\alpha k)}\sin(\alpha\pi k)}
}
\examples{
x <- seq(0,15,0.25)
y <- dTSS(x,0.5,1,0.3)
plot(x,y)

}
\references{
Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distributions and
processes' \url{https://doi.org/10.1016/j.spa.2013.06.012}
}
