% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{Estim_Des_Temp}
\alias{Estim_Des_Temp}
\title{Function title}
\usage{
Estim_Des_Temp(
  TemperedType = c("Classic", "Subordinator"),
  EstimMethod = c("ML", "GMM", "Cgmm", "Kout"),
  ...
)
}
\arguments{
\item{TemperedType}{A String. Either "Classic", "Subordinator", or "Normal".}

\item{EstimMethod}{A String. Either "ML", "GMM", "Cgmm", or "Kout".}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
