% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{initOutputFile}
\alias{initOutputFile}
\title{Function title}
\usage{
initOutputFile(thetaT, MCparam, TemperedType, Estimfct, ...)
}
\arguments{
\item{thetaT}{A gap holder.}

\item{MCparam}{A gap holder.}

\item{TemperedType}{A gap holder.}

\item{Estimfct}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
