% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{m.opt.retst}
\alias{m.opt.retst}
\title{Optimal constant for fast rejection}
\usage{
m.opt.retst(V0)
}
\arguments{
\item{V0}{numeric vector >= 0}
}
\value{
optimal constant m for the fast rejection algorithm
}
\description{
Note: this is mainly to show that this function can be very well
approximated much more simply by just using m <- round(V0).
}
\author{
Martin Maechler (based on Marius Hofert's code)
}
