% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{NameParamsObjectsTemp}
\alias{NameParamsObjectsTemp}
\title{Function title}
\usage{
NameParamsObjectsTemp(mat, type = c("Classic", "Subordinator", "Normal"))
}
\arguments{
\item{mat}{A positive integer.}

\item{type}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
Not sure, if export is really necessary.
}
\details{
Gap holder for details.
}
