% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{parallelizeMCsimulation}
\alias{parallelizeMCsimulation}
\title{Function to parallelize the Monte Carlo Simulation}
\usage{
parallelizeMCsimulation(MCparam = 10000, cores = NULL, ...)
}
\arguments{
\item{MCparam}{Number of Monte Carlo simulation for each couple of parameter,
default=100; integer}

\item{cores}{size of cluster for parallelization. Positive Integer.}
}
\value{
The return object is a list of 2. Results of the simulation are
listed in \code{$outputMat}.
}
\description{
Since the Monte Carlo Simulation is very computationally intensive, it may
be worthwhile to split it across all available processor cores. To do this,
simply pass all the parameters from the \code{\link[=TemperedEstim_Simulation]{TemperedEstim_Simulation()}}
function to this function in the same way.
}
\details{
details
}
