% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pNTS}
\alias{pNTS}
\title{Title}
\usage{
pNTS(
  q,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  a = -40,
  b = 40,
  nf = 2^13,
  ...
)
}
\arguments{
\item{q}{A gap holder.}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{A gap holder}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{a}{A gap holder. -20 by default.}

\item{b}{A gap holder. 20 by default.}

\item{nf}{A gap holder. 2048 by default. I don't know if this param is used
in this function.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
pNTS(0.1,0.5,1,1,1,1)
pNTS(0.1,0.5,1,1,1,1,NULL, -20, 20, 2^6)

}
