% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pSTS}
\alias{pSTS}
\title{Function title}
\usage{
pSTS(
  q,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  pmethod = "integrate",
  N = 10000,
  ...
)
}
\arguments{
\item{q}{A numeric vector of quantiles.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{pmethod}{A gap holder.}

\item{N}{integer: the number of replications, if
\code{pmethod != "integrate"}. 10000 by default.}

\item{...}{A gap holder.}
}
\value{
A gap holder.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
pSTS(3,0.7,1.354,0.3)
pSTS(1,0.5,10,300,NULL,"")

}
