% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pTSS}
\alias{pTSS}
\title{Cumulative probability function of the tempered stable subordinator
distribution}
\usage{
pTSS(
  q,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  pmethod = "integrate",
  N = 8192,
  ...
)
}
\arguments{
\item{q}{A numeric vector of positive quantile.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{pmethod}{A string. If not "integrate", the function \code{chartocdf()}
will be triggered.}

\item{N}{is a power of two & N >= 1024. if
\code{pmethod != "integrate"}. 10000 by default. Relevant for}

\item{...}{Possibility to modify \code{stats::integrate()}.}
}
\value{
As \code{q} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the cumulative probability function of the tempered
stable subordinator distribution.
}
\examples{
\donttest{
x <- seq(0,15,0.25)
y <- pTSS(x,0.7,1.354,0.3)
plot(x,y)
}

}
\seealso{
See also the [dTSS()] density-function.
}
