% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{qCTS}
\alias{qCTS}
\title{Quantile function of the classic tempered stable (CTS)}
\usage{
qCTS(
  p,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  qmin = NULL,
  qmax = NULL,
  ...
)
}
\arguments{
\item{p}{A numeric vector of probabilities. Each probability must be a real
number >0 and <1.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{qmin, qmax}{Limits of the interval. Will be calcultated if
\code{==NULL}.}

\item{...}{Modify [pSTS()] and [stats::uniroot()].}
}
\value{
As \code{p} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the quantile function of the classic tempered stable
(CTS).
}
\examples{
\donttest{
  qCTS(0.5,1.5,10,10,10,10,10)
  qCTS(0.5,1.5,1,1,1,1,1)
}

}
\seealso{
See also the [pCTS()] probability function.
}
