% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{qSTS}
\alias{qSTS}
\title{Quantile function of the tempered stable subordinator distribution}
\usage{
qSTS(
  p,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  qmin = NULL,
  qmax = NULL,
  ...
)
}
\arguments{
\item{p}{A numeric vector of probabilities. Each probability must be a real
number >0 and <1.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{qmin, qmax}{Limits of the interval. Will be calcultated if
\code{==NULL}.}

\item{...}{Modify [pSTS()] and [stats::uniroot()].}
}
\value{
As \code{p} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the quantile function of the tempered stable
subordinator distribution.
}
\examples{
qSTS(0.5,0.5,5,0.01)
qSTS(0.5,0.9,1,10,NULL)

}
\seealso{
See also the [pSTS()] probability function.
}
