% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rCGMY}
\alias{rCGMY}
\title{Function title}
\usage{
rCGMY(n, C, G, M, Y, method = "SR", k = 100, ...)
}
\arguments{
\item{n}{A gap holder.}

\item{C}{A real number > 0.}

\item{G, M}{A  real number > 0.}

\item{Y}{A real number between 0 and 2.}

\item{method}{A String. Either "aAR" or "SR". "SR" by default.}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
rCGMY(100,1,1,1,0.5)
rCGMY(100,1,1,1,0.5, "SR", k= 100)

}
