% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rCTS}
\alias{rCTS}
\title{Function title}
\usage{
rCTS(
  n,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  method = "aAR",
  k = 100,
  c = 1
)
}
\arguments{
\item{n}{A gap holder.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "aAR" or "SR".}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}

\item{c}{A real number. Only relevant for \code{method == "aAR"}.
TODO(Till): Choose a suitable default value.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
rCTS(10,0.5,1,1,1,1,1,NULL,"SR",10)
rCTS(10,0.5,1,1,1,1,1,NULL,"aAR")

}
