% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{retstableC}
\alias{retstableC}
\title{Efficiently sampling an exponentially tilted stable distribution}
\usage{
retstableC(alpha, V0, h = 1, method = NULL)
}
\arguments{
\item{alpha}{parameter in (0,1]}

\item{V0}{vector of random variates}

\item{h}{non-negative real number}

\item{method}{which method to call ("Marius Hofert", "Luc Devroye")}
}
\value{
vector of variates St
}
\description{
Sample random variates St ~ \tilde{S}(alpha, 1, (cos(alpha\emph{pi/2)
\emph{V_0)^{1/alpha}, V_0}I_{alpha = 1}, h}I_{alpha != 1}; 1) with
Laplace-Stieltjes transform exp(-V_0((h+t)^alpha-h^alpha)), see Nolan's book for
the parametrization, with the fast rejection.
This procedure is more efficient than retstableR since it calls the C
function retstable_c and uses both the fast rejection and Luc Devroye's algorithm.
}
\author{
Martin Maechler
}
