% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{retstableR}
\alias{retstableR}
\title{Sampling an exponentially tilted stable distribution}
\usage{
retstableR(alpha, V0, h = 1)
}
\arguments{
\item{alpha}{parameter in (0,1]}

\item{V0}{vector of random variates}

\item{h}{non-negative real number}
}
\value{
vector of variates St
}
\description{
Sample a vector of random variates St ~ \tilde{S}(alpha, 1,
(cos(alpha\emph{pi/2)\emph{V_0)^{1/alpha}, V_0}I_{alpha = 1},
h}I_{alpha != 1}; 1) with LS transform
exp(-V_0((h+t)^alpha-h^alpha)) with the fast rejection
algorithm; see Nolan's book for the parametrization
}
\author{
Marius Hofert, Martin Maechler
}
