% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{retstablerej}
\alias{retstablerej}
\title{Sample an exponentially tilted stable distribution as an m-fold sum}
\usage{
retstablerej(m, V0, alpha)
}
\arguments{
\item{m}{number of summands, any positive integer}

\item{V0}{random variate}

\item{alpha}{parameter in (0,1]}
}
\value{
St
}
\description{
Sample a random variate St ~ \tilde{S}(alpha, 1, (cos(alpha\emph{pi/2)
\emph{V_0)^{1/alpha}, V_0}I_{alpha = 1}, I_{alpha != 1}; 1) with
Laplace-Stieltjes transform exp(-V_0((1+t)^alpha-1)), see Nolan's book for
the parametrization, via an m-fold sum of random variates from
\tilde{S}(alpha, 1, (cos(alpha}pi/2)*V_0/m)^{1/alpha}, (V_0/m)
\emph{I_{alpha = 1}, I_{alpha != 1}; 1) with Laplace-Stieltjes transform
exp(-(V_0/m)}((1+t)^alpha-1)). This is a building block for the fast rejection.
}
\author{
Marius Hofert, Martin Maechler
}
