% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rNTS}
\alias{rNTS}
\title{Function to generate random variates of NTS distribution.}
\usage{
rNTS(
  n,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  method = "AR",
  k = 100
)
}
\arguments{
\item{n}{sample size (integer).}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{A gap holder.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "AR" or "SR". "AR" by default.}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}
}
\value{
generates random deviates.
}
\description{
Generates \code{n} random numbers distributed according to the normal
tempered stable distribution.
}
\details{
TODO: Describe difference between "AR" and "SR". Right now, AR is the method
that is used in Monte Carlo. And it cannot be changed there.
}
\examples{
rNTS(100, 0.5, 1,1,1,1)
rNTS(10, 0.6, 0,1,1,0)
rNTS(10, 0.5, 1,1,1,1, NULL, "SR", 100)

}
