% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{rstable1C}
\alias{rstable1C}
\title{Efficiently sampling stable distributions}
\usage{
rstable1C(n, alpha, beta, gamma = 1, delta = 0, pm = 1)
}
\arguments{
\item{n}{number of random variates to be generated}

\item{alpha, }{see code in fBasics}

\item{beta, }{see code in fBasics}

\item{gamma, }{see code in fBasics}

\item{delta, }{see code in fBasics}

\item{pm}{in {0,1} parameterization, see code in fBasics}
}
\value{
vector of variates S
}
\description{
Sample S ~ S(alpha, beta, gamma, delta; pm), see rstable1R() above.
For beta == 1 and pm == 1, the fast C implementation is used.
}
\author{
Martin Maechler
}
