% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retstable.R
\name{rstable1R}
\alias{rstable1R}
\title{Sampling stable distributions}
\usage{
rstable1R(n, alpha, beta, gamma = 1, delta = 0, pm = 1)
}
\arguments{
\item{n}{number of random variates to be generated}

\item{alpha, }{see code in fBasics}

\item{beta, }{see code in fBasics}

\item{gamma, }{see code in fBasics}

\item{delta, }{see code in fBasics}

\item{pm}{in {0,1} parameterization, see code in fBasics}
}
\value{
vector of variates S
}
\description{
Sample S ~ S(alpha, beta, gamma, delta; pm), see package
\pkg{stabledist} for the parameterization.
}
\author{
Martin Maechler, based on Diethelm Wuertz's code in fBasics
}
