% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rSTS}
\alias{rSTS}
\title{Title}
\usage{
rSTS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 100,
  ...
)
}
\arguments{
\item{n}{A gap holder.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "AR" or "SR".}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
rSTS(100,0.5,1,1)
rSTS(100,0.5,1,1,NULL,"SR",50)

}
