% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{TemperedEstim_Simulation}
\alias{TemperedEstim_Simulation}
\title{Monte Carlo Simulation}
\usage{
TemperedEstim_Simulation(
  ParameterMatrix,
  SampleSizes = c(200, 1600),
  MCparam = 100,
  TemperedType = c("Classic", "Subordinator", "Normal", "CGMY"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  saveOutput = TRUE,
  SeedOptions = NULL,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{ParameterMatrix}{A gap holder.}

\item{SampleSizes}{Sample sizes to be used to simulate the data. By default,
we use 200 (small sample size) and 1600 (large sample size);
vector of integer.}

\item{MCparam}{Number of Monte Carlo simulation for each couple of parameter,
default=100; integer}

\item{TemperedType}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{Estimfct}{The estimation function to be used. A String.
Either "ML", "GMM", "Cgmm", or "GMC".}

\item{HandleError}{Logical flag: if set to TRUE, the simulation doesn't stop
when an error in the estimation function is encountered. A vector of
(size 4) NA is saved and the the simulation carries on. See details.}

\item{saveOutput}{Logical flag: if set to TRUE, a csv file (for each couple
of parameter) with the the estimation
information is saved in the current directory. See details.}

\item{SeedOptions}{List to control the seed generation. See details.}

\item{eps}{A gap holder.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Runs Monte Carlo simulation for a selected estimation method. The function
can save results in a file.
}
\details{
\strong{Error Handling} It is advisable to set it to TRUE when user is
planning to launch long simulations as it will prevent the procedure to stop
if an error occurs for one sample data. The estimation function will produce
a vector of NA as estimated parameters related to this (error generating)
sample data and move on to the next Monte Carlo step.

\strong{Output file} Setting \code{saveOutput} to \code{TRUE} will have the
side effect of saving a csv file in the working directory. This file will
have \code{MCparam*length(SampleSizes)} lines and its columns will be:
\describe{
  \item{alphaT, ...:}{the true value of the parameters.}
  \item{data size:}{the sample size used to generate the simulated data.}
  \item{seed:}{the seed value used to generate the simulated data.}
  \item{alphaE, ...:}{the estimate of the 4 parameters.}
  \item{failure:}{binary: 0 for success, 1 for failure.}
  \item{time:}{estimation running time in seconds.}
}
The file name is informative to let the user identify the value of the true
parameters, the MC parameters as well as the options selected for the
estimation method. The csv file is updated after each MC estimation which is
useful when the simulation stops before it finishes.

\strong{SeedOptions} If user does not want to control the seed generation,
he could ignore this argument (default value NULL). This argument can be
more useful when one wants to cut the simulation (even for one parameter
value) into pieces. In that case, he can control which part of the seed
vector he wants to use.
\describe{
  \item{MCtot:}{total values of MC simulations in the entire process.}
  \item{seedStart:}{starting index in the seed vector. The vector extracted
  will be of size MCparam.}
}
}
\examples{
TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.5,1,1,1,1,0),
                                                c(0.5,1,1,1,1,0)),
                        SampleSizes = 10, MCparam = 10,
                        TemperedType = "Classic", Estimfct = "ML",
                        saveOutput = FALSE)

}
\seealso{
\url{https://github.com/GeoBosh/StableEstim/blob/master/R/Simulation.R}
}
