% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{TemperedEstim_v2}
\alias{TemperedEstim_v2}
\title{Function title}
\usage{
TemperedEstim_v2(
  TemperedType = c("Classic", "Subordinator", "Normal", "CGMY"),
  EstimMethod = c("ML", "GMM", "Cgmm", "GMC"),
  data,
  theta0 = NULL,
  ComputeCov = FALSE,
  HandleError = TRUE,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{TemperedType}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{EstimMethod}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{data}{A gap holder.}

\item{theta0}{A gap holder. \code{NULL} by default.}

\item{ComputeCov}{A Boolean. \code{FALSE} by default.}

\item{HandleError}{A Boolean. \code{TRUE} by default.}

\item{eps}{A gap holder. \code{1e-06} by default.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
