% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempStable.R
\docType{package}
\name{TempStable}
\alias{TempStable}
\title{TempStable: Describe package in one sentence}
\description{
Here is more space to go into detail.
Users can enter this documentation with ?TempStable
}
\section{Basic Functions}{

Here you can use sections to divide content.
}

