% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charCTS}
\alias{charCTS}
\title{Title}
\usage{
charCTS(
  t,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A gap holder.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
charCTS(2,0.5,1,1.5,2,2.5,3)
charCTS(2,1.5,200,150,300,2500,1,NULL)

}
