% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS_extended.R
\name{charGTS}
\alias{charGTS}
\title{Characteristic function of the generalized classical tempered stable
distribution}
\usage{
charGTS(
  t,
  alphap = NULL,
  alpham = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A vector of real numbers where the CF is evaluated.}

\item{alphap, alpham}{Stability parameter. A real number between 0 and 1.}

\item{deltap, deltam}{Scale parameter. A real number > 0.}

\item{lambdap, lambdam}{Tempering parameter. A real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{Parameters stacked as a vector.}
}
\value{
The CF of the the generalized classical tempered stable distribution.
}
\description{
gapholder
}
\details{
gapholder
}
\examples{
gapholder

}
\references{
gapholder
}
