% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS_extended.R
\name{charKRTS}
\alias{charKRTS}
\title{Characteristic function of the Kim-Rachev tempered stable distribution}
\usage{
charKRTS(
  t,
  alpha = NULL,
  kp = NULL,
  km = NULL,
  rp = NULL,
  rm = NULL,
  pp = NULL,
  pm = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A vector of real numbers where the CF is evaluated.}

\item{alpha}{Stability parameter. A real number between 0 and 1.}

\item{kp, km}{Gapholder.}

\item{rp, rm}{Gapholder.}

\item{pp, pm}{Gapholder.}

\item{mu}{A location parameter, any real number.}

\item{theta}{Parameters stacked as a vector.}
}
\value{
The CF of the the Kim-Rachev tempered stable distribution.
}
\description{
gapholder
}
\details{
Gapholder. Currently different function with "functionOrigin" is no option,
as the second is not working.
}
\examples{
gapholder

}
\references{
gapholder
}
