% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS_extended.R
\name{charMTS}
\alias{charMTS}
\title{Characteristic function of the modified tempered stable distribution}
\usage{
charMTS(
  t,
  alpha = NULL,
  delta = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  functionOrigin = "kim09"
)
}
\arguments{
\item{t}{A vector of real numbers where the CF is evaluated.}

\item{alpha}{Stability parameter. A real number between 0 and 1.}

\item{delta}{Scale parameter. A real number > 0.}

\item{lambdap, lambdam}{Tempering parameter. A real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{Parameters stacked as a vector.}

\item{functionOrigin}{A string. Either "kim09", or "kim08".}
}
\value{
The CF of the the modified tempered stable distribution.
}
\description{
gapholder
}
\details{
\strong{Origin of functions}
\describe{
\item{kim09}{Ansatz aus: Kim et al. 2009 The modified tempered stable
distribution, GARCH-models and option pricing. Alpha darf hier von -Inf
bis 1 gehen. Ausgenommen wird alpha == 1/2.}
\item{kim08}{Ansatz aus: Kim et al. 2008? Financial market models with
Levy processes and time-varying volatility.}
}
}
\examples{
gapholder

}
\references{
gapholder
}
