% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{dCTS}
\alias{dCTS}
\title{Density function of the classic tempered stable (CTS) distribution}
\usage{
dCTS(
  x,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  dens_method = "FFT",
  a = -20,
  b = 20,
  nf = 2048
)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{dens_method}{Algorithm for numerical evaluation. Choose between \code{
"FFT"} and \code{"Conv"}.}

\item{a}{Starting point of FFT, if \code{dens_method == "FFT"}. -20
by default.}

\item{b}{Ending point of FFT, if \code{dens_method == "FFT"}. 20
by default.}

\item{nf}{Pieces the transformation is divided in. Limited to power-of-two
size.}
}
\value{
As \code{x} is a numeric vector, the return value is also a numeric
vector.
}
\description{
The probability density function of tempered stable distributions is
generally not available in closed form. Crucially, even a simple relationship
with a stable density as for the TSS distribution is not available.
For numerical evaluations it is therefore necessary to rely on algorithms
like the fast Fourier transform applied to the characteristic function.
}
\examples{
dCTS(1,0.6,1,1,1,1,1,NULL,"FFT",-20,20,2048)
dCTS(1,0.6,1,1,1,1,1,NULL,"Conv")

}
\seealso{
Massing, Till (2022), 'Parametric Estimation of Tempered Stbale Laws'
}
