% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{parallelizeMCsimulation}
\alias{parallelizeMCsimulation}
\title{title}
\usage{
parallelizeMCsimulation(
  ParameterMatrix,
  SampleSizes = c(1000),
  MCparam = 10000,
  TemperedType = c("Classic", "Subordinator", "Normal"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  eps = 1e-06,
  cores = NULL,
  ...
)
}
\arguments{
\item{ParameterMatrix}{The matrix is to be composed of vectors, row by row.
Each vector must fit the pattern of theta of the \code{TemperedType}.}

\item{SampleSizes}{Sample sizes to be used to simulate the data. By default,
we use 200 (small sample size) and 1600 (large sample size);
vector of integer.}

\item{MCparam}{Number of Monte Carlo simulation for each couple of parameter,
default=100; integer}

\item{TemperedType}{A String. Either "Classic", "Subordinator", or "Normal".}

\item{Estimfct}{The estimation function to be used. A String.
Either "ML", "GMM", "Cgmm", or "GMC".}

\item{HandleError}{Logical flag: if set to TRUE, the simulation doesn't stop
when an error in the estimation function is encountered. A vector of
(size 4) NA is saved and the the simulation carries on. See details..}

\item{eps}{Numerical error tolerance. \code{1e-06} by default.}

\item{cores}{size of cluster for parallelization. Positive Integer.}
}
\value{
The return object is a list of 2. Results of the simulation are
listed in \code{$outputMat}.
}
\description{
description
}
\details{
details
}
