% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pCTS}
\alias{pCTS}
\title{Cumulative probability function of the classic tempered stable (CTS)
distribution}
\usage{
pCTS(
  q,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  a = -40,
  b = 40,
  nf = 2^13,
  ...
)
}
\arguments{
\item{q}{A numeric vector of quantile.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{a}{Starting point of FFT, if \code{dens_method == "FFT"}. -20
by default.}

\item{b}{Ending point of FFT, if \code{dens_method == "FFT"}. 20
by default.}

\item{nf}{Pieces the transformation is divided in. Limited to power-of-two
size.}

\item{...}{Possibility to modify \code{stats::integrate()}.}
}
\value{
As \code{q} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the cumulative probability function of the CTS
distribution.
}
\examples{
x <- seq(-5,5,0.25)
y <- pCTS(x,0.5,1,1,1,1,1)
plot(x,y)

}
\seealso{
See also the [dCTS()] density-function.
}
