% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pSTS}
\alias{pSTS}
\title{Cumulative probability function of the tempered stable subordinator
distribution}
\usage{
pSTS(
  q,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  pmethod = "integrate",
  N = 10000,
  ...
)
}
\arguments{
\item{q}{A numeric vector of positive quantile.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{pmethod}{A string. If not "integrate", the function \code{chartocdf()}
will be triggered.}

\item{N}{integer: the number of replications, if
\code{pmethod != "integrate"}. 10000 by default.}

\item{...}{Possibility to modify \code{stats::integrate()}.}
}
\value{
As \code{x} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the cumulative probability function of the tempered
stable suborinator distribution.
}
\examples{
x <- seq(-15,15,0.25)
y <- pSTS(x,0.7,1.354,0.3)
plot(x,y)

}
\seealso{
See also the [dSTS()] density-function.
}
