% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{qNTS}
\alias{qNTS}
\title{Quantile function of the normal tempered stable (NTS)}
\usage{
qNTS(
  p,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL,
  qmin = NULL,
  qmax = NULL,
  ...
)
}
\arguments{
\item{p}{A numeric vector of probabilities. Each probability must be a real
number >0 and <1.}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{A gap holder.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number >= 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{qmin, qmax}{Limits of the interval. Will be calcultated if
\code{==NULL}.}

\item{...}{Modify [pSTS()] and [stats::uniroot()].}
}
\value{
As \code{p} is a numeric vector, the return value is also a numeric
vector.
}
\description{
This function returns the quantile function of the normal tempered stable
(NTS).
}
\examples{
qNTS(0.1,0.5,1,1,1,1)
qNTS(0.3,0.6,1,1,1,1,NULL)

}
\seealso{
See also the [pSTS()] probability function.
}
