% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{qSTS}
\alias{qSTS}
\title{Quantile function of of the tempered stable subordinator distribution}
\usage{
qSTS(
  p,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  qmin = NULL,
  qmax = NULL,
  ...
)
}
\arguments{
\item{p}{A gap holder.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{qmin, qmax}{Limits of the interval.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
TODO: die Quantilfunktion schaut, an welcher x -Stelle p = pSTS() ist
}
\details{
Gap holder for details.
}
\examples{
qSTS(0.5,0.5,5,0.01)
qSTS(0.5,0.9,1,10,NULL)

}
