% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rCTS}
\alias{rCTS}
\title{Function to generate random variates of CTS distribution.}
\usage{
rCTS(
  n,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  method = "aAR",
  k = 100,
  c = 1
)
}
\arguments{
\item{n}{sample size (integer).}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdam}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "aAR" or "SR".}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}

\item{c}{A real number. Only relevant for \code{method == "aAR"}.}
}
\value{
generates random deviates.
}
\description{
Generates \code{n} random numbers distributed according to the classic
tempered stable (CTS) distribution.
}
\details{
"aAR" stands for the approximate Acceptance-Rejection Method and "SR" for
infinite shot noise series representation. "aAR" is the standard method used
in Monte Carlo simulation. For more details, see references.
}
\examples{
rCTS(10,0.5,1,1,1,1,1,NULL,"SR",10)
rCTS(10,0.5,1,1,1,1,1,NULL,"aAR")

}
\references{
Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random
variates' \doi{10.1016/j.cam.2010.12.014}
}
