% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rTSS}
\alias{rTSS}
\title{Function to generate random variates of TSS distribution.}
\usage{
rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 100
)
}
\arguments{
\item{n}{sample size (integer).}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "AR" or "SR".}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}
}
\value{
generates random deviates.
}
\description{
Generates \code{n} random numbers distributed according to the distribution
of the tempered stable subordinator.
}
\details{
"AR" stands for the Acceptance-Rejection Method and "SR" for infinite shot
noise series representation. "AR" is the standard method used in Monte Carlo
simulation. For more details, see references.
}
\examples{
rTSS(100,0.5,1,1)
rTSS(100,0.5,1,1,NULL,"SR",50)

}
\references{
Massing, Till (2022), 'Parametric Estimation of Tempered Stable Laws'

Kawai, R & Masuda, H (2011), 'On simulation of tempered stable random
variates' \url{https://doi.org/10.1016/j.cam.2010.12.014}
}
