% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{rTSS}
\alias{rTSS}
\title{Function to generate random variates of TSS distribution.}
\usage{
rTSS(
  n,
  alpha = NULL,
  delta = NULL,
  lambda = NULL,
  theta = NULL,
  method = "AR",
  k = 100
)
}
\arguments{
\item{n}{sample size (integer).}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}

\item{method}{A String. Either "AR" or "SR".}

\item{k}{integer: the number of replications, if \code{method == "SR"}. 100
by default.}
}
\value{
generates random deviates.
}
\description{
Generates \code{n} random numbers distributed according to the distribution
of the tempered stable subordinator.
}
\details{
TODO: Describe difference between "AR" and "SR". Right now, AR is the method
that is used in Monte Carlo. And it cannot be changed there.
}
\examples{
rTSS(100,0.5,1,1)
rTSS(100,0.5,1,1,NULL,"SR",50)

}
