% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{TemperedEstim}
\alias{TemperedEstim}
\title{Main estimation function}
\usage{
TemperedEstim(
  TemperedType = c("Classic", "Subordinator", "Normal"),
  EstimMethod = c("ML", "GMM", "Cgmm", "GMC"),
  data,
  theta0 = NULL,
  ComputeCov = FALSE,
  HandleError = TRUE,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{TemperedType}{A String. Either "Classic", "Subordinator", or "Normal"}

\item{EstimMethod}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{data}{Data used to perform the estimation: vector of length n.}

\item{theta0}{A vector of numeric values corresponding to the pattern of the
\code{TemperedType}.}

\item{ComputeCov}{Logical flag: If set to TRUE, the asymptotic covariance
matrix is computed. \code{FALSE} by default.}

\item{HandleError}{Logical flag: If set to \code{TRUE} and if an error occurs
during the estimation procedure, the computation will carry on and NA will be
returned. Useful for Monte Carlo simulations.\code{TRUE} by default.}

\item{eps}{A gap holder. \code{1e-06} by default.}

\item{...}{Other arguments to be passed to the estimation function or the
asymptotic confidence level.}
}
\value{
Object of a estim-class. See details for more information.
}
\description{
Main estimation function which computes all the information about the
estimator. It allows the user to choose the preferred method and several
related options.
}
\details{
\strong{TemperedType} Detailed documentation of the individual tempered
stable distributions can be viewed in the respective characteristic function.
Use \code{charTSS}, \code{charCTS}, or \code{charNTS}.

\strong{Estimfct} Detailed documentation of the individual parameters
of the various estimate functions is available in the package
\code{StableEstim}.
\describe{
  \item{For ML:}{use \code{?StableEstim::MLParametersEstim}. See usage of
  Maximum likelihood estimation in Kim et al. (2008)}
  \item{For GMM:}{use \code{?StableEstim::GMMParametersEstim}. Generalized
  Method of Moments by Hansen (1982)}
  \item{For Cgmm:}{use \code{?StableEstim::CgmmParametersEstim}. Continuum
  Generalized Methods of Moments by Carrasco & Kotchoni (2017)}
  \item{For GMC:}{We also use a method of moment approach which follows
   Kuechler & Tappe (2013). They match empirical cumulants with their
   theoretical counterparts. We extend this by using Hansen's (1982) GMM
   framework. We call the approach generalized method of cumulants (GMC) to
   distinguish it from the GMM method using characteristic function moment
   conditions. However, it fits well into Hansen's (1982) framework allowing
   for standard asymptotic theory.}
}

\strong{Estim-Class} Class storing all the information about the estimation
method; output of this function.
\strong{Slots of the return class}
\describe{
  \item{par:}{Object of class "\code{numeric}"; Value of the estimated
  parameters.}
  \item{par0:}{Object of class "\code{numeric}";Initial guess for the
  parameters.}
  \item{vcov:}{Object of class "\code{matrix}" representing the covariance
  matrix.}
  \item{confint:}{Object of class "\code{matrix}" representing the confidence
  interval computed at a specific level (attribute of the object).}
  \item{data:}{Object of class "\code{numeric}" used to compute the
  estimation.}
  \item{sampleSize:}{Object of class "\code{numeric}" ; length of the data.}
  \item{others:}{Object of class "\code{list}" ; more information about the
  estimation method.}
  \item{duration:}{Object of class "\code{numeric}" ; duration in seconds.}
  \item{failure:}{Object of class "\code{numeric}" representing the status of
   the procedure: 0 failure or 1 success.}
  \item{method:}{Object of class "\code{character}" description of the
  parameter used in the estimation.}
}
}
\examples{
\donttest{
TemperedEstim(TemperedType = "Classic", EstimMethod = "ML",
               data = rCTS(2,1.5,1,1,1,1,0),
               theta0 = c(1.5,1,1,1,1,0) - 0.1);
TemperedEstim("Subordinator", "GMM", rTSS(20,0.5,1,1), algo = "2SGMM",
              alphaReg = 0.01, regularization = "cut-off",
              WeightingMatrix = "OptAsym", t_scheme = "free",
              t_free = seq(0.1,2,length.out = 12));
TemperedEstim("Normal", "Cgmm", rNTS(20,0.5,1,1,1,0), algo = "2SCgmm",
              alphaReg = 0.01, subdivisions = 20,
              IntegrationMethod = "Uniform", randomIntegrationLaw = "unif",
              s_min = 0, s_max= 1);
TemperedEstim("Subordinator", "GMC", rTSS(20, 0.5, 1, 1), algo = "2SGMC",
              alphaReg = 0.01, WeightingMatrix = "OptAsym",
              regularization = "cut-off", ncond = 8);
}
}
\references{
Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws';

Kim, Y. s., Rachev, S. T., Bianchi, M. L. & Fabozzi, F. J. (2008), 'Financial
market models with lévy processes and time-varying volatility'
\doi{10.1016/j.jbankfin.2007.11.004};

Hansen, L. P. (1982), 'Large sample properties of generalized method of
moments estimators' \doi{10.2307/1912775};

Carrasco, M. & Kotchoni, R. (2017), 'Efficient estimation using the
characteristic function' \doi{10.1017/S0266466616000025};

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distribution and processes'
\doi{10.1016/j.spa.2013.06.012};
}
\seealso{
\url{https://github.com/GeoBosh/StableEstim/blob/master/R/Simulation.R}
}
