% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{TemperedEstim}
\alias{TemperedEstim}
\title{Function title}
\usage{
TemperedEstim(
  TemperedType = c("Classic", "Subordinator", "Normal"),
  EstimMethod = c("ML", "GMM", "Cgmm", "GMC"),
  data,
  theta0 = NULL,
  ComputeCov = FALSE,
  HandleError = TRUE,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{TemperedType}{A String. Either "Classic", "Subordinator", or "Normal"}

\item{EstimMethod}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{data}{A gap holder.}

\item{theta0}{A gap holder. \code{NULL} by default. Must not be NULL, if
data comes from function like \code{rCTS(...)}}

\item{ComputeCov}{A Boolean. \code{FALSE} by default.}

\item{HandleError}{A Boolean. \code{TRUE} by default.}

\item{eps}{A gap holder. \code{1e-06} by default.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
\strong{Estimfct} Detailed documentation of the individual parameters
of the various estimate functions is available in the package
\code{StableEstim}.
\describe{
\item{For ML:}{Write \code{?StableEstim::MLParametersEstim} to console.}
\item{For GMM:}{Write \code{?StableEstim::GMMParametersEstim} to console.}
\item{For Cgmm:}{Write \code{?StableEstim::CgmmParametersEstim} to console.}
\item{For GMC:}{TODO}
}
}
\examples{
\donttest{
TemperedEstim(TemperedType = "Classic", EstimMethod = "ML",
               data = rCTS(2,1.5,1,1,1,1,0),
               theta0 = c(1.5,1,1,1,1,0) - 0.1);
TemperedEstim("Subordinator", "GMM", rSTS(20,0.5,1,1), algo = "2SGMM",
              alphaReg = 0.01, regularization = "cut-off",
              WeightingMatrix = "OptAsym", t_scheme = "free",
              t_free = seq(0.1,2,length.out = 12));
TemperedEstim("Normal", "Cgmm", rNTS(20,0.5,1,1,1,0), algo = "2SCgmm",
              alphaReg = 0.01, subdivisions = 20,
              IntegrationMethod = "Uniform", randomIntegrationLaw = "unif",
              s_min = 0, s_max= 1);
TemperedEstim("Subordinator", "GMC", rSTS(20, 0.5, 1, 1), algo = "2SGMC",
              alphaReg = 0.01, WeightingMatrix = "OptAsym",
              regularization = "cut-off", ncond = 8);
}
}
