% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{TemperedEstim}
\alias{TemperedEstim}
\title{Function title}
\usage{
TemperedEstim(
  TemperedType = c("Classic", "Subordinator", "Normal", "CGMY"),
  EstimMethod = c("ML", "GMM", "Cgmm", "GMC"),
  data,
  theta0 = NULL,
  ComputeCov = FALSE,
  HandleError = TRUE,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{TemperedType}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{EstimMethod}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{data}{A gap holder.}

\item{theta0}{A gap holder. \code{NULL} by default.}

\item{ComputeCov}{A Boolean. \code{FALSE} by default.}

\item{HandleError}{A Boolean. \code{TRUE} by default.}

\item{eps}{A gap holder. \code{1e-06} by default.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
testData <- c(1.8873152, -0.4843727,  0.4755897, -0.1257814,  1.3484823,
             -0.3866821, -0.4258380, -0.4658479, -2.9774961,  0.9646364,
             -0.5875601, -2.0316790,  0.3641900,  1.1882307,  1.6635770,
             -0.0554876,  0.4005471,  0.7820444, -0.3786902,  1.5131663)
TemperedEstim("Classic","ML",testData)

}
