% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{TemperedEstim_Simulation}
\alias{TemperedEstim_Simulation}
\title{Monte Carlo Simulation}
\usage{
TemperedEstim_Simulation(
  ParameterMatrix,
  SampleSizes = c(200, 1600),
  MCparam = 100,
  TemperedType = c("Classic", "Subordinator", "Normal"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  saveOutput = TRUE,
  SeedOptions = NULL,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{ParameterMatrix}{A gap holder.}

\item{SampleSizes}{Sample sizes to be used to simulate the data. By default,
we use 200 (small sample size) and 1600 (large sample size);
vector of integer.}

\item{MCparam}{Number of Monte Carlo simulation for each couple of parameter,
default=100; integer}

\item{TemperedType}{A String. Either "Classic", "Subordinator", or "Normal".}

\item{Estimfct}{The estimation function to be used. A String.
Either "ML", "GMM", "Cgmm", or "GMC".}

\item{HandleError}{Logical flag: if set to TRUE, the simulation doesn't stop
when an error in the estimation function is encountered. A vector of
(size 4) NA is saved and the the simulation carries on. See details.}

\item{saveOutput}{Logical flag: if set to TRUE, a csv file (for each couple
of parameter) with the the estimation
information is saved in the current directory. See details.}

\item{SeedOptions}{List to control the seed generation. See details.}

\item{eps}{A gap holder.}

\item{...}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Runs Monte Carlo simulation for a selected estimation method. The function
can save results in a file.
}
\details{
\strong{Error Handling} It is advisable to set it to TRUE when user is
planning to launch long simulations as it will prevent the procedure to stop
if an error occurs for one sample data. The estimation function will produce
a vector of NA as estimated parameters related to this (error generating)
sample data and move on to the next Monte Carlo step.

\strong{Output file} Setting \code{saveOutput} to \code{TRUE} will have the
side effect of saving a csv file in the working directory. This file will
have \code{MCparam*length(SampleSizes)} lines and its columns will be:
\describe{
  \item{alphaT, ...:}{the true value of the parameters.}
  \item{data size:}{the sample size used to generate the simulated data.}
  \item{seed:}{the seed value used to generate the simulated data.}
  \item{alphaE, ...:}{the estimate of the parameters.}
  \item{failure:}{binary: 0 for success, 1 for failure.}
  \item{time:}{estimation running time in seconds.}
}
The file name is informative to let the user identify the value of the true
parameters, the MC parameters as well as the options selected for the
estimation method. The csv file is updated after each MC estimation which is
useful when the simulation stops before it finishes.

\strong{SeedOptions} If user does not want to control the seed generation,
he could ignore this argument (default value NULL). This argument can be
more useful when one wants to cut the simulation (even for one parameter
value) into pieces. In that case, he can control which part of the seed
vector he wants to use.
\describe{
  \item{MCtot:}{total values of MC simulations in the entire process.}
  \item{seedStart:}{starting index in the seed vector. The vector extracted
  will be of size MCparam.}
}

\strong{Estimfct} Detailed documentation of the individual parameters
of the various estimate functions is available in the package
\code{StableEstim}.
\describe{
  \item{For ML:}{use \code{?StableEstim::MLParametersEstim}. See usage of
  Maximum likelihood estimation in Kim et al. (2008)}
  \item{For GMM:}{use \code{?StableEstim::GMMParametersEstim}. Generalized
  Method of Moments by Hansen (1982)}
  \item{For Cgmm:}{use \code{?StableEstim::CgmmParametersEstim}. Continuum
  Generalized Methods of Moments by Carrasco & Kotchoni (2017)}
  \item{For GMC:}{We also use a method of moment approach which follows
   Kuechler & Tappe (2013). They match empirical cumulants with their
   theoretical counterparts. We extend this by using Hansen's (1982) GMM
   framework. We call the approach generalized method of cumulants (GMC) to
   distinguish it from the GMM method using characteristic function moment
   conditions. However, it fits well into Hansen's (1982) framework allowing
   for standard asymptotic theory.}
}
}
\examples{
\donttest{
TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.5,1,1,1,1,0),
                                                 c(0.5,1,1,1,1,0)),
                         SampleSizes = 10, MCparam = 10,
                         TemperedType = "Classic", Estimfct = "ML",
                         saveOutput = FALSE)

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.5,1,1,1,1,0)),
                         SampleSizes = 40, MCparam = 40,
                         TemperedType = "Classic", Estimfct = "GMM",
                         saveOutput = FALSE, algo = "2SGMM",
                         regularization = "cut-off",
                         WeightingMatrix = "OptAsym", t_scheme = "free",
                         alphaReg = 0.005,
                         t_free = seq(0.1,2,length.out=12))

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.45,0.55,1,1,1,0)),
                         SampleSizes = 4, MCparam = 4,
                         TemperedType = "Classic", Estimfct = "Cgmm",
                         saveOutput = FALSE, algo = "2SCgmm",
                         alphaReg = 0.01, subdivisions = 20,
                         IntegrationMethod = "Uniform",
                         randomIntegrationLaw = "unif",
                         s_min = 0, s_max= 1)

TemperedEstim_Simulation(ParameterMatrix = rbind(c(1.45,0.55,1,1,1,0)),
                         SampleSizes = 4, MCparam = 4,
                         TemperedType = "Classic", Estimfct = "GMC",
                         saveOutput = FALSE, algo = "2SGMC",
                         alphaReg = 0.01, WeightingMatrix = "OptAsym",
                         regularization = "cut-off", ncond = 8)
}

}
\references{
Massing, T. (2022), 'Parametric Estimation of Tempered Stable Laws';

Kim, Y. s., Rachev, S. T., Bianchi, M. L. & Fabozzi, F. J. (2008), 'Financial
market models with lévy processes and time-varying volatility'
\url{https://doi.org/10.1016/j.jbankfin.2007.11.004};

Hansen, L. P. (1982), 'Large sample properties of generalized method of
moments estimators' \url{https://doi.org/10.2307/1912775};

Carrasco, M. & Kotchoni, R. (2017), 'Efficient estimation using the
characteristic function' \url{https://doi.org/10.1017/S0266466616000025};

Kuechler, U. & Tappe, S. (2013), 'Tempered stable distribution and processes'
\url{https://doi.org/10.1016/j.spa.2013.06.012};
}
\seealso{
\url{https://github.com/GeoBosh/StableEstim/blob/master/R/Simulation.R}
}
