% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{TemperedEstim_Simulation}
\alias{TemperedEstim_Simulation}
\title{Function title}
\usage{
TemperedEstim_Simulation(
  ParameterMatrix,
  SampleSizes = c(200, 1600),
  MCparam = 100,
  TemperedType = c("Classic", "Subordinator", "Normal", "CGMY"),
  Estimfct = c("ML", "GMM", "Cgmm", "GMC"),
  HandleError = TRUE,
  FctsToApply = StatFcts,
  saveOutput = TRUE,
  StatSummary = FALSE,
  CheckMat = TRUE,
  tolFailCheck = tolFailure,
  SeedOptions = NULL,
  eps,
  ...
)
}
\arguments{
\item{ParameterMatrix}{A gap holder.}

\item{SampleSizes}{A gap holder.}

\item{MCparam}{A gap holder.}

\item{TemperedType}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{Estimfct}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}

\item{HandleError}{A Boolean. \code{TRUE} by default.}

\item{FctsToApply}{A gap holder.}

\item{saveOutput}{A gap holder.}

\item{StatSummary}{A gap holder.}

\item{CheckMat}{A gap holder.}

\item{tolFailCheck}{A gap holder.}

\item{SeedOptions}{A gap holder.}

\item{eps}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
